pr comment fixes.
