+++ timeclock/urls.py
@@ -17,7 +17,7 @@ from django.conf.urls import include, url
 from django.contrib import admin
 
 urlpatterns = [
-	url(r'^accounts/', include('accounts.urls')),
+    url(r'^accounts/', include('accounts.urls')),
     url(r'^admin/', include(admin.site.urls)),
     url('^', include('django.contrib.auth.urls')),
 ]
